feature #377 : ShareView 분리
